컨텐츠 스타일 개선
