+++ src/css/content.css
@@ -1,22 +1,22 @@
 @charset "utf-8";
 
 .content {
-  padding: 20px 0 40px;
   word-wrap: break-word;
   color: #333;
   font-size: 14px;
   margin: 0;
+  font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
 }
 
 .content h1 {
-  font-size: 28px;
+  font-size: 24px;
   font-weight: 700;
   margin: 24px 0 12px 0;
   line-height: 1.3;
 }
 
 .content h2 {
-  font-size: 22px;
+  font-size: 20px;
   font-weight: 700;
   margin: 20px 0 10px 0;
   line-height: 1.35;
@@ -34,16 +34,6 @@
 .content h6 {
   font-size: 1em;
   margin: 2em 0 1em
-}
-
-.content h1,
-.content h2,
-.content h3,
-.content h4,
-.content h5,
-.content h6 {
-  font-family: sans-serif;
-  font-weight: 400;
 }
 
 .content p {
@@ -137,7 +127,8 @@ ol li {
 
 .content img {
   max-width: 100%;
-  height: auto
+  height: auto;
+  border-radius: 12px;
 }
 
 .content .image-group {
